chore: split main code for easier testing
